Refactor: 도메인 모듈을 중심으로 의존성 역전을 적용하도록 모듈 구조 수정
