+++ module-api/src/main/kotlin/finn/moduleApi/service/PredictionQueryService.kt
@@ -1,21 +1,25 @@
 package finn.moduleApi.service
 
-import finn.moduleApi.paging.PageResponse
 import finn.moduleApi.paging.PredictionPageRequest
-import finn.moduleApi.queryDto.PredictionDetailQueryDto
-import finn.moduleApi.repository.PredictionRepository
 import finn.moduleDomain.entity.Prediction
+import finn.moduleDomain.paging.PageResponse
+import finn.moduleDomain.queryDto.PredictionDetailQueryDto
+import finn.moduleDomain.repository.PredictionRepository
 import org.springframework.stereotype.Service
 import java.util.*
 
 @Service
 class PredictionQueryService(private val predictionRepository: PredictionRepository) {
 
-    fun getPredictionList(pageRequest: PredictionPageRequest) : PageResponse<Prediction> {
-        return predictionRepository.getPredictionList(pageRequest)
+    fun getPredictionList(pageRequest: PredictionPageRequest): PageResponse<Prediction> {
+        return predictionRepository.getPredictionList(
+            pageRequest.page,
+            pageRequest.size,
+            pageRequest.sort
+        )
     }
 
-    fun getPredictionDetail(tickerId: UUID) : PredictionDetailQueryDto {
+    fun getPredictionDetail(tickerId: UUID): PredictionDetailQueryDto {
         return predictionRepository.getPredictionDetail(tickerId)
     }
 }
