+++ module-api/src/main/kotlin/finn/moduleApi/service/TickerQueryService.kt
@@ -1,7 +1,7 @@
 package finn.moduleApi.service
 
-import finn.moduleApi.queryDto.TickerSearchQueryDto
-import finn.moduleApi.repository.TickerRepository
+import finn.moduleDomain.queryDto.TickerSearchQueryDto
+import finn.moduleDomain.repository.TickerRepository
 import org.springframework.stereotype.Service
 
 @Service
